snippetgenerator: start snippet with best matching text extract
Until now, a snippet contained all text extracts, that contain at
least one of the query terms. As a consequence, text extracts with
a high coverage of search terms could show up late in the snippet.

This patch updates the snippet generation to always start with the
best matching extract in the text, followed by any extracts which
match the query just as well.